Feat-5 move components tests to correct location, add test for LocationButton
